requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.23.5 